add ErrorStateModel file
