+++ CriticalMass/ErrorStateModel.swift
@@ -1,0 +1,19 @@
+//
+//  ErrorStateModel.swift
+//  CriticalMaps
+//
+//  Created by Malte Bünz on 04.11.19.
+//  Copyright © 2019 Pokus Labs. All rights reserved.
+//
+
+import Foundation
+
+struct ErrorStateModel {
+    let errorTitle: String
+    let errorMessage: String
+}
+
+extension ErrorStateModel {
+    static let `default` = ErrorStateModel(errorTitle: String.errorStateTitle,
+                                           errorMessage: String.errorStateMessage)
+}
